requirements.txt
fastapi
uvicorn[standard]
pymongo
requests
python-dotenv
google-generativeai